Infra: new tests have been added. Feature: Added the ability to disable one of the models (so far only daytime) using the flag

+++ SolSignalModel1D_Backtest.Core/Data/DataBuilder/DataRow.cs
@@ -16,7 +16,7 @@
 		public double BtcRet1 { get; set; }
 		public double BtcRet3 { get; set; }
 
-		public int Fng { get; set; }
+		public double Fng { get; set; }
 		public double DxyChg30 { get; set; }
 		public double GoldChg30 { get; set; }
 		public double BtcVs200 { get; set; }
